run-rdma: Add check and mount of configfs
It looks like docker has issues when using symbolic links inside
configfs when it is mounted from inside a container. Ensure configfs
is mounted before the container spins up. We cannot do it in the
script because we use the docker user group for that and it is not
desirable to make that root.